misc refactoring and import map generation edge case handling
