+++ packages/cli/src/lib/node-modules-utils.js
@@ -33,7 +33,37 @@ async function getPackageJsonForProject({ userWorkspace, projectDirectory }) {
       : {};
 }
 
+function mergeImportMap(html = '', map = {}, shouldShim = false) {
+  const importMapType = shouldShim ? 'importmap-shim' : 'importmap';
+  const hasImportMap = html.indexOf(`script type="${importMapType}"`) > 0;
+  const danglingComma = hasImportMap ? ',' : '';
+  const importMap = JSON.stringify(map, null, 2).replace('}', '').replace('{', '');
+
+  if (Object.entries(map).length === 0) {
+    return html;
+  }
+
+  if (hasImportMap) {
+    return html.replace('"imports": {', `
+      "imports": {
+        ${importMap}${danglingComma}
+    `);
+  } else {
+    return html.replace('<head>', `
+      <head>
+      <script type="${importMapType}">
+        {
+          "imports": {
+            ${importMap}
+          }
+        }
+      </script>
+    `);
+  }
+}
+
 export {
   getPackageJsonForProject,
-  getResolvedHrefFromPathnameShortcut
+  getResolvedHrefFromPathnameShortcut,
+  mergeImportMap
 };
