+++ packages/plugin-graphql/src/index.js
@@ -1,6 +1,6 @@
 import fs from 'fs/promises';
 import { graphqlServer } from './core/server.js';
-import { mergeImportMap } from '@greenwood/cli/src/lib/walker-package-ranger.js';
+import { mergeImportMap } from '@greenwood/cli/src/lib/node-modules-utils.js';
 import { ResourceInterface } from '@greenwood/cli/src/lib/resource-interface.js';
 import { ServerInterface } from '@greenwood/cli/src/lib/server-interface.js';
